Pop datastore tests out to shared file
